move initial runner to it's own file
